app/requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
selenium==4.15.2